util functions in seperate file
