--- console-frontend/src/utils/double.js
@@ -1,3 +1,0 @@
-export default function double(x) {
-  return 2 * x;
-}
